+++ packages/interop/test/fixtures/create-helia.ts
@@ -1,10 +1,11 @@
-import { createHelia, HeliaInit } from 'helia'
-import { createLibp2p } from 'libp2p'
-import { tcp } from '@libp2p/tcp'
 import { noise } from '@chainsafe/libp2p-noise'
 import { yamux } from '@chainsafe/libp2p-yamux'
+import { tcp } from '@libp2p/tcp'
 import { MemoryBlockstore } from 'blockstore-core'
 import { MemoryDatastore } from 'datastore-core'
+import { createHelia, type HeliaInit } from 'helia'
+import { createLibp2p } from 'libp2p'
+import { identifyService } from 'libp2p/identify'
 import type { Helia } from '@helia/interface'
 
 export async function createHeliaNode (init?: Partial<HeliaInit>): Promise<Helia> {
@@ -27,8 +28,8 @@ export async function createHeliaNode (init?: Partial<HeliaInit>): Promise<Helia
       yamux()
     ],
     datastore,
-    nat: {
-      enabled: false
+    services: {
+      identify: identifyService()
     }
   })