+++ packages/interface/CHANGELOG.md
@@ -1,3 +1,53 @@
+## [@helia/interface-v1.2.2](https://github.com/ipfs/helia/compare/@helia/interface-v1.2.1...@helia/interface-v1.2.2) (2023-08-05)
+
+
+### Trivial Changes
+
+* update project config ([#175](https://github.com/ipfs/helia/issues/175)) ([f185a72](https://github.com/ipfs/helia/commit/f185a7220a62f7fc0c025aa5c0be5a981c63cc48))
+
+
+### Dependencies
+
+* **dev:** bump aegir from 39.0.13 to 40.0.8 ([#198](https://github.com/ipfs/helia/issues/198)) ([4d75ecf](https://github.com/ipfs/helia/commit/4d75ecffb79e5177da35d3106e42dac7bc63153a))
+
+## [@helia/interface-v1.2.1](https://github.com/ipfs/helia/compare/@helia/interface-v1.2.0...@helia/interface-v1.2.1) (2023-06-07)
+
+
+### Bug Fixes
+
+* pass options to blockstore.get during pin.add ([#148](https://github.com/ipfs/helia/issues/148)) ([3a5234e](https://github.com/ipfs/helia/commit/3a5234e3c2f88f9910678b0cbbac5fd340117cc9))
+
+## [@helia/interface-v1.2.0](https://github.com/ipfs/helia/compare/@helia/interface-v1.1.1...@helia/interface-v1.2.0) (2023-06-07)
+
+
+### Features
+
+* add offline option to blockstore get ([#145](https://github.com/ipfs/helia/issues/145)) ([71c5f6b](https://github.com/ipfs/helia/commit/71c5f6bc32b324ee237e56c2c5a1ce903b3bdbef))
+
+
+### Trivial Changes
+
+* update changelogs ([#142](https://github.com/ipfs/helia/issues/142)) ([fefd374](https://github.com/ipfs/helia/commit/fefd3744c0a6d8471de31762ece6ec59b65496c1))
+
+## [@helia/interface-v1.1.1](https://github.com/ipfs/helia/compare/@helia/interface-v1.1.0...@helia/interface-v1.1.1) (2023-05-19)
+
+
+### Bug Fixes
+
+* add helia version to agent version ([#128](https://github.com/ipfs/helia/issues/128)) ([48e19ec](https://github.com/ipfs/helia/commit/48e19ec545cc67157e14ae59054fa377a583cb01)), closes [#122](https://github.com/ipfs/helia/issues/122)
+
+## [@helia/interface-v1.1.0](https://github.com/ipfs/helia/compare/@helia/interface-v1.0.0...@helia/interface-v1.1.0) (2023-05-19)
+
+
+### Features
+
+* provide default libp2p instance ([#127](https://github.com/ipfs/helia/issues/127)) ([45c9d89](https://github.com/ipfs/helia/commit/45c9d896afa27f5ea043cc5f576d50fc4fa556e9)), closes [#121](https://github.com/ipfs/helia/issues/121)
+
+
+### Trivial Changes
+
+* bump aegir from 38.1.8 to 39.0.4 ([#111](https://github.com/ipfs/helia/issues/111)) ([2156568](https://github.com/ipfs/helia/commit/215656870cb821dd6be2f8054dc39932ba25af14))
+
 ## @helia/interface-v1.0.0 (2023-03-23)
 
 
@@ -31,9 +81,7 @@
 * update publish config ([913ab6a](https://github.com/ipfs/helia/commit/913ab6ae9a2970c4b908de04b8b6a236b931a3b0))
 * update release please config ([b52d5e3](https://github.com/ipfs/helia/commit/b52d5e3eecce41b10640426c339c99ad14ce874e))
 
-# Changelog
-
-## [0.1.0](https://www.github.com/ipfs/helia/compare/interface-v0.0.0...interface-v0.1.0) (2023-03-22)
+## @helia/interface-v0.1.0 (2023-03-22)
 
 
 ### Features